[MCUX-78594][MCXE31]Fix MSG issue in SIUL2 driver

+++ MCXE/MCXE31B/drivers/fsl_siul2.c
@@ -187,7 +187,7 @@ static uint16_t SIUL2_Reverse_Bit_16(uint16_t value)
 
     for (i = 0U; i < 8U; i++)
     {
-        temp = (((value >> i) & 1U) << (15U - i)) | (((value << i) & 0x8000U) >> (15U - i));
+        temp = ((((uint32_t)value >> i) & 1U) << (15U - i)) | ((((uint32_t)value << i) & 0x8000U) >> (15U - i));
         ret |= (uint16_t)(temp & 0xFFFFU);
     }
